Se reemplazan los tipos de datos long por int

+++ tp-orgDeDatos/Pruebas/Algoritmos de compresion/HashMap.cpp
@@ -19,29 +19,29 @@ int HashMap::fhash(char* clave){
 	return clave[0]*255+clave[1];
 }
 
-list<unsigned long> HashMap::get(char first, char second){
+list<unsigned int> HashMap::get(char first, char second){
     char* key = (char*) malloc (sizeof(char)*2);
     key[0] = first;
     key[1] = second;
-	list<unsigned long> lista;
+	list<unsigned int> lista;
 	int hash = fhash(key);
 	if (!table[hash])
 		return lista;
 	return table[hash]->getValue();
 }
 
-void HashMap::put(char first, char second, unsigned long value){
+void HashMap::put(char first, char second, unsigned int value){
     char* key = (char*) malloc (sizeof(char)*2);
     key[0] = first;
     key[1] = second;
-	list <unsigned long> lista;
+	list <unsigned int> lista;
 	int hash = fhash(key);
 	if(!table[hash]){
 		lista.push_front(value);
 		table[hash] = new HashEntry(key, lista);
 	}
 	else{
-		list <unsigned long> lista = table[hash]->getValue();
+		list <unsigned int> lista = table[hash]->getValue();
 		lista.push_front(value);
 		table[hash] = new HashEntry(key, lista);
 	}
