+++ tp-orgDeDatos/Pruebas/Compresor/Compresor.h
@@ -23,13 +23,13 @@ class Compresor{
 		 * Va a llamar a SymbolRanking comprimir. El resultado de esto se lo pasa a
 		 * Estructurado.comprimir y modifica/devuelve el archivo comprimido
 		 */
-		void comprimir(char* buffer, unsigned long size);
+		void comprimir(char* buffer, unsigned int size);
 
 
 		/* A partir del buffer obtenido, se genera el archivo descomprimido siguiendo los pasos y logica inversa del
 		 * metodo comprimir.
 		 */
-		void descomprimir(unsigned short* entrada, unsigned long size);
+		void descomprimir(unsigned short* entrada, unsigned int size);
 };
 
 #endif
